feat(readers): implement readImageFile
readImageFile passes the file data as an ArrayBuffer to a web worker in the transfer list. Inside
the web worker, the file's contents are mounted to the Emscripten filesystem with the WORKERFS as a
Blob. The generated Emscripten modules are no longer UMD modules to permit proper loading in the web
worker.

BREAKING CHANGE: Emscripten modules no longer wrapped as UMD modules.