feat(itkConfig): ITK module configuration specified with itkConfig.itkModulesPath
This provides a single path to configure the related module paths that are dynamically loaded.

The web workers, Emscripten modules, and Emscripten WebAssembly modules are
all located relative to each other, so specify the value in a single configuration
location.

Use ES6 import syntax for browser-only modules.

BREAKING CHANGE: itkConfig is now aliased in Webpack configuration with './itkConfig$' and the
modules path are specified with itkConfig.itkModulesPath.