feat(BrowserAPI): Return and accept the web worker used for computation
This enables re-use of the web worker or calling `terminate()` on the worker to free resources. We
also fix a bug -- previously only a single worker was being used by the
app.

BREAKING CHANGE: Calls to the browser API must now pass null or a previously generated web worker as
their first argument. The browser API function now return objects with a `webWorker` member.