fix(worker): Ensure worker promises are reused
When running task in a worker in "parallel" the access to the
worker must be performed through the same worker promise otherwise
requests and responses can get interleaved. This situation can occur
if for example Promise.all(...) is used to run the tasks in a
worker. To prevent this issue we attach the work promise to the
worker, so it can be used again if the worker is reused.