build(DynamicImports): Add @vite-ignore
We do not want vite/rollup to try to bundle yet per

  https://github.com/InsightSoftwareConsortium/itk-wasm/issues/476

Addresses:

```
4:44:15 PM [vite] warning:
/home/matt/src/itk-wasm/dist/core/internal/loadEmscriptenModuleMainThread.js
21 |      const wasmBinary = response.data;
22 |      const fullModulePath = `${modulePrefix}.js`;
23 |      const result = await import(/* webpackIgnore: true */ fullModulePath);
   |                                  ^
24 |      const instantiated = result.default({ wasmBinary });
25 |      return instantiated;
The above dynamic import cannot be analyzed by vite.
See https://github.com/rollup/plugins/tree/master/packages/dynamic-import-vars#limitations
for supported dynamic import formats. If this is intended to be left as-is, you can use the
/* @vite-ignore */ comment inside the import() call to suppress this warning.
```