feat(KOS): Key object selection (KOS) structured report (SR)
Bump DCMTK to include changes for fixing dcmsr. This will
include the fix for accepting the hyperlink prefix for
key object references as an argument (std::string).

$ git shortlog 66fce68..f9da0c6
Shreeraj Jadhav (1):
      ENH: Replace HTML_HYPERLINK_PREFIX_FOR_CGI with std::string argument

Modify operation structured-report-to-html to accept the hyperlink
prefix as a string argument. This will be passed to the
DSRDocument::renderHTML() function.

Generate/update TS bindings for structured-report-to-html.