feat(JsonObject): Add JSON_OUTPUT as output type
Add JSON_OUTPUT as a possible output type which will be generated
by an operation (.cxx) as a string, but will be parsed as a
generic Javascript object when receiving results on the JS
side of the pipeline.

Use JSON_OUTPUT for the second output of applyPresentationStateToImage operation.

Add convenience command 'npm run bindgen' to package.json.