feat: add gaussian-kernel-radius pipeline
May differ in some cases from the current corresponding ngff-zarr Python
function because the constrainedMaxKernelWidth is only limited by the
size in the current direction as opposed to all directions.

We are just using an VDimension template argument to GaussianOperator
of 1 always because this does not look to be used in the result.